Move squid code into its own file
Also de-dupe the logic between egress and ingress. The existing code had
two separate functions which were nearly identical.
